App Flow Change
Application now loads the splash screen first, then from the splash screen can point to any desired page.
Should be Login / Register
