+++ ui/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"connectivity","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/connectivity-0.4.6+2/","dependencies":[]},{"name":"path_provider","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider-1.6.0/","dependencies":[]},{"name":"shared_preferences","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences-0.5.6+1/","dependencies":[]}],"android":[{"name":"connectivity","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/connectivity-0.4.6+2/","dependencies":[]},{"name":"path_provider","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider-1.6.0/","dependencies":[]},{"name":"shared_preferences","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences-0.5.6+1/","dependencies":[]}],"macos":[{"name":"shared_preferences_macos","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_macos-0.0.1+4/","dependencies":[]}],"linux":[],"windows":[],"web":[{"name":"shared_preferences_web","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_web-0.1.2+3/","dependencies":[]}]},"dependencyGraph":[{"name":"connectivity","dependencies":[]},{"name":"path_provider","dependencies":[]},{"name":"shared_preferences","dependencies":["shared_preferences_macos","shared_preferences_web"]},{"name":"shared_preferences_macos","dependencies":[]},{"name":"shared_preferences_web","dependencies":[]}],"date_created":"2020-02-05 01:47:38.891544","version":"1.14.6"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"connectivity","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/connectivity-0.4.6+2/","dependencies":[]},{"name":"path_provider","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider-1.6.0/","dependencies":[]},{"name":"shared_preferences","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences-0.5.6+1/","dependencies":[]}],"android":[{"name":"connectivity","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/connectivity-0.4.6+2/","dependencies":[]},{"name":"path_provider","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider-1.6.0/","dependencies":[]},{"name":"shared_preferences","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences-0.5.6+1/","dependencies":[]}],"macos":[{"name":"shared_preferences_macos","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_macos-0.0.1+4/","dependencies":[]}],"linux":[],"windows":[],"web":[{"name":"shared_preferences_web","path":"/home/morgan/Development/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_web-0.1.2+3/","dependencies":[]}]},"dependencyGraph":[{"name":"connectivity","dependencies":[]},{"name":"path_provider","dependencies":[]},{"name":"shared_preferences","dependencies":["shared_preferences_macos","shared_preferences_web"]},{"name":"shared_preferences_macos","dependencies":[]},{"name":"shared_preferences_web","dependencies":[]}],"date_created":"2020-02-05 17:21:19.941862","version":"1.14.6"}
